Rename coinPriceHandler
