+++ src/index.ts
@@ -8,7 +8,7 @@ import { handler as vestedBalanceHandler } from './handlers/vestedBalance';
 import { handler as delegatorCountHandler } from './handlers/delegatorCount';
 import { handler as totalDelegatorsHandler } from './handlers/totalDelegators';
 import { handler as totalStakedCoinsHandler } from './handlers/totalStakedCoins';
-import { handler as coinPriceHandler } from './handlers/coinPrice';
+import { handler as allArbitrageOpportunitiesHandler } from './handlers/allArbitrageOpportunities';
 import { handler as arbitrageOpportunitiesHandler } from './handlers/arbitrageOpportunities';
 import { webhookTriggers } from './handlers/webhookTriggers';
 
@@ -24,12 +24,12 @@ addEventListener('fetch', (event: FetchEvent) => {
 
 function registerRoutes(router: Router) {
   router.get('/', totalSupplyHandler);
-  router.get('/arbitrage/all', coinPriceHandler);
   router.get('/arbitrage', arbitrageOpportunitiesHandler);
-  router.get('/balances/total/:address', totalBalanceHandler);
+  router.get('/arbitrage/all', allArbitrageOpportunitiesHandler);
   router.get('/balances/liquid/:address', liquidBalanceHandler);
-  router.get('/balances/vesting/:address', vestingBalanceHandler);
+  router.get('/balances/total/:address', totalBalanceHandler);
   router.get('/balances/vested/:address', vestedBalanceHandler);
+  router.get('/balances/vesting/:address', vestingBalanceHandler);
   router.get('/staking/delegators/total', totalDelegatorsHandler);
   router.get('/staking/delegators/:validator_address', delegatorCountHandler);
   router.get('/supply/circulating', circulatingSupplyHandler);
